LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/math-script-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"

    <!-- speculative case in document.write -->
    <math><script src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=3cbcd322-9507-4457-9e59-d08ea6b2719f&amp;encodingcheck=&Gbreve;"></script></math>


FAIL Speculative parsing, document.write(): math-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"
